+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.4] - 2019-02-19
+### ADDED
+- Added `AbstractRequestApi` In Super
+- Required `GuzzleHttp ^6.3` in Composer
+
+## [0.4.3] - 2019-02-19
+### FIXED
+- Fixed Namespace Error of PlatformBaseModel
+
 ## [0.4.2] - 2019-02-19
 ### ADDED
 - Copied PlatformBaseModel to Base